extract sidebar component
