+++ client/web-sveltekit/src/lib/search/CodeHostIcon.svelte
